refactor: clarify store topographer helper subpaths (#467)
## Context

Store adapter-binding helpers and Topographer direct DB/admin helpers were living on package roots, blurring the v1 public boundary.

## What Changed

- Added `@ontrails/store/adapter-support` for adapter-facing store binding helpers.
- Added `@ontrails/topographer/backend-support` for direct database/admin helpers.
- Updated Drizzle, Trails app code, Warden tests, package READMEs, and API docs to the new canonical subpaths.
- Removed root-barrel exceptions from Warden's public internal import guard.

## How To Test

- `bun run --cwd packages/store typecheck`
- `bun run --cwd packages/topographer typecheck`
- `bun test packages/store packages/topographer`
- `bun run --cwd adapters/drizzle typecheck`
- `bun test adapters/drizzle`
- Final stack-tip verification: `bun run typecheck`, `bun run test`, `bun run dead-code`, `bun run publish:check`

## Risks / Rollout Notes

This is a pre-v1 breaking root-export cleanup for helper APIs. The new subpaths are explicit and documented.

## Release / Changeset

Includes a major changeset for `@ontrails/store` and `@ontrails/topographer`, plus patch entries for first-party packages that consume the new subpaths.

Closes: TRL-644
