+++ .changeset/public-helper-subpaths.md
@@ -1,0 +1,10 @@
+---
+"@ontrails/store": major
+"@ontrails/topographer": major
+"@ontrails/core": patch
+"@ontrails/drizzle": patch
+"@ontrails/trails": patch
+"@ontrails/warden": patch
+---
+
+Move store adapter-binding helpers to `@ontrails/store/adapter-support` and topographer direct database/admin helpers to `@ontrails/topographer/backend-support`, keeping root exports focused on contract-level APIs.
