+++ packages/topographer/src/index.ts
@@ -83,14 +83,3 @@ export type {
   TopoStoreTrailDetailRecord,
   TopoStoreTrailRecord,
 } from './topo-store.js';
-export {
-  countPinnedSnapshots,
-  countPrunableSnapshots,
-  countTopoSnapshots,
-  pruneUnpinnedSnapshots,
-} from './internal/topo-snapshots.js';
-export {
-  createTopoSnapshot as createStoredTopoSnapshot,
-  getStoredTopoExport,
-} from './internal/topo-store.js';
-export type { StoredTopoExport } from './internal/topo-store.js';
