+++ packages/warden/src/__tests__/drift.test.ts
@@ -11,12 +11,12 @@ import {
   Result,
 } from '@ontrails/core';
 import {
-  createStoredTopoSnapshot,
   createTopoStore,
   deriveSurfaceMapHash,
   deriveSurfaceMap,
   writeSurfaceLock,
 } from '@ontrails/topographer';
+import { createStoredTopoSnapshot } from '@ontrails/topographer/backend-support';
 import { z } from 'zod';
 
 import { checkDrift } from '../drift.js';
