+++ packages/warden/src/rules/public-internal-deep-imports.ts
@@ -21,9 +21,6 @@ import type { WardenPublicWorkspace } from '../workspaces.js';
 const RULE_NAME = 'public-internal-deep-imports';
 const ONTRAILS_SPECIFIER_PATTERN = /^(@ontrails\/[^/]+)(?:\/(.+))?$/;
 const ROOT_BARREL_INTERNAL_RE_EXPORT_ALLOWLIST = new Set([
-  '@ontrails/store:./internal/signal-identity.js',
-  '@ontrails/topographer:./internal/topo-snapshots.js',
-  '@ontrails/topographer:./internal/topo-store.js',
   '@ontrails/tracing:./internal/dev-state.js',
 ]);
